+++ crates/rpc/src/eth/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod pubsub;
 pub(crate) mod rpc;
+pub(crate) mod types;
